# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Landevo Real Estate Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threads:
    get:
      operationId: listThreads
      tags: [messages]
      summary: List message threads for current user
      responses:
        "200":
          description: Threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageThread"

    post:
      operationId: createThread
      tags: [messages]
      summary: Start a new message thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageThreadInput"
      responses:
        "201":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageThread"
  /threads/{threadId}/messages:
    get:
      operationId: getThreadMessages
      tags: [messages]
      summary: Get messages in a thread
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a thread
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Notifications ─────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - in: query
          name: unreadOnly
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageParticipant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
      required: [id, name, role]
    MessageThread:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: ["integer", "null"]
        listingTitle:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/MessageParticipant"
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, participants, unreadCount, createdAt]
    MessageThreadInput:
      type: object
      properties:
        recipientId:
          type: integer
        listingId:
          type: integer
        initialMessage:
          type: string
          minLength: 1
      required: [recipientId, initialMessage]
    Message:
      type: object
      properties:
        id:
          type: integer
        threadId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        readAt:
          type: ["string", "null"]
          format: date-time
      required: [id, threadId, senderId, senderName, content, createdAt]
    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]

    # ── Storage paths are added at the end of paths section ───────────────

    # ── Notifications ─────────────────────────────────────────────────────
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [offer_received, offer_accepted, offer_rejected, listing_verified, listing_rejected, transaction_update, new_message, system]
        title:
          type: string
        body:
          type: string
        readAt:
          type: ["string", "null"]
          format: date-time
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, body, createdAt]

    # ── Dashboard ─────────────────────────────────────────────────────────
...